+++ cw/code/ui/fixtures.py
@@ -4,6 +4,8 @@ from webdriver_manager.chrome import ChromeDriverManager
 from webdriver_manager.firefox import GeckoDriverManager
 from ui.pages.base_page import BasePage
 from selenium.webdriver.chrome.options import Options
+from selenium.webdriver.chrome.service import Service as ChromeService
+from selenium.webdriver.firefox.service import Service as FirefoxService
 
 
 def _options():
@@ -16,10 +18,13 @@ def _options():
 def get_driver(browser_name):
     if browser_name == 'chrome':
         browser = webdriver.Chrome(
-            executable_path=ChromeDriverManager().install(), options=_options())
+            service=ChromeService(
+                executable_path=ChromeDriverManager().install()),
+            options=_options())
     elif browser_name == 'firefox':
         browser = webdriver.Firefox(
-            executable_path=GeckoDriverManager().install())
+            service=FirefoxService(
+                executable_path=GeckoDriverManager().install()))
     else:
         raise RuntimeError(f'Unsupported browser: "{browser_name}"')
     browser.maximize_window()
