+++ cw/code/ui/locators/base_locators.py
@@ -1,0 +1,5 @@
+from selenium.webdriver.common.by import By
+
+
+class BaseLocators():
+    GET_LI_ELEMENT = (By.TAG_NAME, 'li')
